+++ R/RcppExports.R
@@ -13,3 +13,35 @@ mars_backward_only_cpp <- function(x_in, y_in, dirs_in, cuts_in, penalty, nprune
     .Call(`_roadrunner_mars_backward_only_cpp`, x_in, y_in, dirs_in, cuts_in, penalty, nprune, nthreads, force_size, return_path, weights_in, R_in, Qty_in)
 }
 
+krls_kernel_cpp <- function(X, sigma) {
+    .Call(`_roadrunner_krls_kernel_cpp`, X, sigma)
+}
+
+krls_kernel_pred_cpp <- function(Xnew, Xtrain, sigma) {
+    .Call(`_roadrunner_krls_kernel_pred_cpp`, Xnew, Xtrain, sigma)
+}
+
+krls_eig_cpp <- function(K) {
+    .Call(`_roadrunner_krls_eig_cpp`, K)
+}
+
+krls_solve_cpp <- function(d, V, Vsq, Vty, lambda) {
+    .Call(`_roadrunner_krls_solve_cpp`, d, V, Vsq, Vty, lambda)
+}
+
+krls_loo_loss_cpp <- function(d, V, Vsq, Vty, lambda) {
+    .Call(`_roadrunner_krls_loo_loss_cpp`, d, V, Vsq, Vty, lambda)
+}
+
+krls_deriv_cpp <- function(X, K, c, sigma) {
+    .Call(`_roadrunner_krls_deriv_cpp`, X, K, c, sigma)
+}
+
+krls_avg_deriv_var_cpp <- function(X, K, V, d, sigma, lambda, sigma2) {
+    .Call(`_roadrunner_krls_avg_deriv_var_cpp`, X, K, V, d, sigma, lambda, sigma2)
+}
+
+krls_vsq_cpp <- function(V) {
+    .Call(`_roadrunner_krls_vsq_cpp`, V)
+}
+